requirements.txt
streamlit>=1.32.0
sqlmodel>=0.0.16
sqlalchemy>=2.0.0
aiohttp>=3.9.0
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
